API Reference

Corleone

Corleone.Trajectory — Type

struct Trajectory{S, U, P, T, SH}

The solution to an optimal control problem.

Fields

sys: The symbolic system used for SymbolicIndexingInterface

u: The state trajectory

p: The parameter values

t: The timepoints

shooting: The shooting values

shooting_indices: The shooting indices

Note

If present, shooting_points contains a list of Tuples (timeseries_index, last_shooting_point).

Corleone.ControlParameter — Type

struct ControlParameter{T, C, B}

Implements a piecewise constant control discretization.

Fields

name: The name of the control

t: The timepoints at which discretized variables are introduced

controls: The initial values for the controls. Either a vector or a function (rng,t,bounds) -> u

bounds: The bounds as a tuple

Corleone.SingleShootingLayer — Type

struct SingleShootingLayer{P, A, C, B, PB, SI, PI} <: LuxCore.AbstractLuxLayer

Defines a callable layer that integrates the AbstractDEProblem problem using the specified algorithm. Controls are assumed to impact differential equation via its parameters problem.p at the positions indicated via control_indices and are itself specified via controls. Moreover, initial conditions problem.u0 that are degrees of freedom to be optimized can be specified by their indices via tunable_ic along with their upper and lower bounds via bounds_ic.

Fields

problem: The underlying differential equation problem

algorithm: The algorithm with which problem is integrated.

control_indices: Indices in parameters of prob corresponding to controls

controls: The controls

tunable_ic: Indices of prob.u0 which are degrees of freedom

bounds_ic: Bounds on the tunable initial conditions of the problem

state_initialization: Initialization of u

tunable_p: Indices of prob.p which are degrees of freedom. This is derived from control_indices!

bounds_p: Bounds on the tunable parameters of the problem

parameter_initialization: Initialization of p

quadrature_indices: Indices of differential states that are quadratures, i.e. they do not enter into the right hand side of problem

Note: The orders of both controls and control_indices, and bounds_ic and tunable_ic are assumed to be identical!

Corleone.MultipleShootingLayer — Type

struct MultipleShootingLayer{L, I, E, Z} <: LuxCore.AbstractLuxWrapperLayer{:layer}

Defines a callable layer that integrates a differential equation using multiple shooting, i.e., the problem is lifted and integration is decoupled on disjunct time intervals given in shooting_intervals. Initial conditions on the shooting_intervals are degrees of freedom (except perhaps for the first layer), for which the initialization scheme initialization provides initial values. Parallelization is integration is possible, for which a suitable EnsembleAlgorithm can be specified with ensemble_alg.

Fields

layer: The original layer

shooting_intervals: The shooting intervals

ensemble_alg: The ensemble algorithm

initialization: The initialization scheme

Corleone.default_initialization — Function

default_initialization

Initializes all shooting nodes with their default value, i.e., their initial value in the underlying problem.

Corleone.random_initialization — Function

random_initialization(rng, shooting; ps, kwargs...)

Initializes all shooting nodes with random values.

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

Keyworded Arguments

ps the default parameters of the shooting layer.

Corleone.forward_initialization — Function

forward_initialization(
    rng,
    shooting;
    ps,
    fixed_indices,
    kwargs...
)

Initializes the problem using a forward solve of the problem. This results in a continuous trajectory.

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

Keyworded Arguments

ps the default parameters of the shooting layer.

Corleone.linear_initialization — Function

linear_initialization(
    rng,
    shooting;
    ps,
    u_infinity,
    kwargs...
)

Linearly interpolates u0 and u_inf for t with tspan[1] < t < tspan[2].

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

Keyworded Arguments

ps the default parameters of the shooting layer.

u_infinity::AbstractArray the value at the last timepoint for all states.

Corleone.custom_initialization — Function

custom_initialization(rng, shooting; ps, u0s, kwargs...)

Initializes all shooting nodes with user-provided values. Initial values are given as an iterable collection of AbstractArrays, Dicts, or a vector of Pairs. The variable indices are interpreted depending on the passed value

AbstractArray just sets the initial condition with the provided value. This assumes equal dimensionality.

Dict assumes the keys represent indices of the initial condition.

A Vector or Tuple of Pairs assumes the first value represents an index of the initial condition.

Other options simply skip the corresponding interval.

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

Keyworded Arguments

ps the default parameters of the shooting layer.

u0s the collection of initial conditions.

Example

# Assumes a shooting layer with 3 intervals
# Skips the first interval
# Sets the first index of the second interval to 3.0
# Sets the second index of the third interval to 5.0 and the 5th index to -1.0
custom_initialization(rng, shooting, u0s = [nothing, Dict(1 => 3.0), (2 => 5.0, 5 => -1.0)])

Corleone.constant_initialization — Function

constant_initialization(rng, shooting; ps, u0, kwargs...)

Initializes all shooting nodes with user-provided value. The variable indices are interpreted depending on the passed value

AbstractArray just sets the initial condition with the provided value. This assumes equal dimensionality.

Dict assumes the keys represent indices of the initial condition.

A Vector or Tuple of Pairs assumes the first value represents an index of the initial condition.

Other options simply skip the corresponding interval.

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

Keyworded Arguments

ps the default parameters of the shooting layer.

u0 the initial condition.

Example

# Assumes a shooting layer with 3 intervals
# Sets the first index of all intervals to 3.0
constant_initialization(rng, shooting, u0 = Dict(1 => 3.0))
constant_initialization(rng, shooting, u0 = (1 => 3.0,))
# Set the initial condition to the given vector
constant_initialization(rng, shooting, u0 = [1., 2., 4.])

Corleone.hybrid_initialization — Function

hybrid_initialization(rng, shooting, f; ps, kwargs...)

Initializes the shooting nodes in a hybrid method by applying the provided methods and indices f sequentially. Here we assume the structure interval => method for the initialization.

Arguments

rng::Random.AbstractRNG a random number generator

shooting::MultipleShootingLayer a shooting layer

f::Pairs of the interval index and applied method

Keyworded Arguments

ps the default parameters of the shooting layer.

All other keyworded arguments are passed on to the functions below.

Corleone.CorleoneDynamicOptProblem — Type

struct CorleoneDynamicOptProblem{L, G, O, C, CB}

A struct for capturing the internal definition of a dynamic optimization problem.

Fields

layer: The resulting layer for the problem

getters: The getters which return the values of the trajectory

objective: The objective function

constraints: The constraint function

lcons: Lower bounds for the constraints

ucons: Upper bounds for the constraints

CorleoneOED

CorleoneOED.OEDLayer — Type

struct OEDLayer{DISCRETE, SAMPLED, FIXED, L, O} <: LuxCore.AbstractLuxWrapperLayer{:layer}

Wraps a Corleone shooting layer with the augmented dynamics needed for optimal experimental design. The DISCRETE type parameter selects discrete or continuous information accumulation, and optional measurement controls can restrict where observations are collected.

Fields

layer: The underlying layer

observed: The observed functions

sampling_indices: The sampling indices

CorleoneOED.fisher_information — Function

fisher_information(oed, x, ps, st; add_initial)

Computes the Fisher information matrix for an OEDLayer at parameters ps and state st.

Arguments

oed: Optimal experimental design layer.

x: External input passed through the Lux layer interface.

ps: Layer parameters, typically from LuxCore.initialparameters or LuxCore.setup.

st: Layer state, typically from LuxCore.initialstates or LuxCore.setup.

Keywords

add_initial: Include previously accumulated information stored in st.

Returns

A tuple (F, st) containing the Fisher information matrix and updated layer state.

CorleoneOED.observed_equations — Function

observed_equations(oed, traj)

Evaluates the observation equations associated with an OEDLayer.

Returns

For trajectory inputs, observed values are returned directly. For layer inputs, the return value is (observed, st) with the updated layer state.

CorleoneOED.sensitivities — Function

sensitivities(oed, traj)

Returns the parameter sensitivities of the observed outputs for an OED trajectory.

Returns

For trajectory inputs, the sensitivity arrays are returned directly. For layer inputs, the return value is (sensitivities, st) with the updated layer state.

CorleoneOED.local_information_gain — Function

local_information_gain(oed, x, ps, st)

Computes local information gain contributions from each observed quantity along the OED trajectory.

Returns

A tuple (gains, st) containing per-observation information matrices and the updated layer state.

CorleoneOED.global_information_gain — Function

global_information_gain(oed, x, ps, st)

Computes global information gain contributions scaled by the inverse final Fisher information matrix.

Returns

A tuple (gains, st) containing per-observation information matrices and the updated layer state.

CorleoneOED.MultiExperimentLayer — Type

struct MultiExperimentLayer{DISCRETE, FIXED, SPLIT, SHOOTING, L, P} <: LuxCore.AbstractLuxLayer

Generalization of OEDLayer to multiple experiments that can be jointly optimized.

Fields

layers: Layers defining multiexperiments

n_exp: Number of experiments

params: Parameters considered in the different experiments

CorleoneOED.ACriterion — Type

struct ACriterion <: CorleoneOED.AbstractCriterion

Implements the ACriterion, i.e., $\textrm{tr}(F^{-1})$.

CorleoneOED.DCriterion — Type

struct DCriterion <: CorleoneOED.AbstractCriterion

Implements the DCriterion, i.e., $\det(F^{-1})$.

CorleoneOED.ECriterion — Type

struct ECriterion <: CorleoneOED.AbstractCriterion

Implements the ECriterion, i.e., $\max\{\lambda: \lambda \textrm{ is eigenvalue of } F^{-1}\}$.

CorleoneOED.FisherACriterion — Type

struct FisherACriterion <: CorleoneOED.AbstractCriterion

Implements the FisherACriterion, i.e., $-\textrm{tr}(F)$

CorleoneOED.FisherECriterion — Type

struct FisherECriterion <: CorleoneOED.AbstractCriterion

Implements the FisherECriterion, i.e., -$\min\{\lambda: \lambda \textrm{ is eigenvalue of } F\}$.

CorleoneOED.FisherDCriterion — Type

struct FisherDCriterion <: CorleoneOED.AbstractCriterion

Implements the FisherDCriterion, i.e., -$\det(F)$.

OptimalControlBenchmarks

OptimalControlBenchmarks.load_benchmarks — Function

load_benchmarks()

Loads all benchmark problem constructors from the problems directory.

Returns

A vector of benchmark constructor functions. Each function accepts a benchmark grid configuration and returns the corresponding optimal control problem data.

OptimalControlBenchmarks.run_all — Function

run_all(benchmarks, optimizer, grids)

Runs each benchmark problem with the Corleone benchmark solver.

Arguments

benchmarks: Iterable of benchmark constructor functions, such as the result of load_benchmarks().

optimizer: Optimization solver passed to solve_with_corleone.

grids: Grid configuration passed to each benchmark constructor.

Returns

A vector of named tuples containing each benchmark name and measured runtime placeholder.